2020-12-21 Make Will Module
DONE_LIST:
All of the questions to make will are saved in questions/question.js and Use WillComponent/* to make the will.

Problem Solved:
When using Drawer navigator, navigation state was saved.
Which means, when you click back and click next, the next component data was saved.
To solve this, I convert createDrawerNavigator into createStackNavigator

TODO:
make sub drawerNavigator for the screens which requires openDrawer